FS_libsで、eigen_FSのバージョンを示す関数を分離した。
